Add a '#' key to send a message to the trash.
Currently, Astroid has no support for deleting messages. Without
deleting messages, a mail server with limited storage will eventually
become full and incapable of receiving any more mail. Message deletion
functionality in clients is an important mechanism for preventing this.

This adds a mechanism to mark messages as 'trash' in Astroid, which
servers like gmail are configured to interpret as a deletion request.